fix name of PlaylistModal & refactor internal types
