+++ client/components/containers/modal/CreatePlaylistModal.vue
@@ -8,7 +8,7 @@
 
 <script lang="ts">
 import { defineComponent, computed } from '@vue/composition-api';
-import PlaylistModal, { On as OnModal, INPUT, Handler } from '~/components/containers/modal/PaylistModal.vue';
+import PlaylistModal, { On as OnModal, INPUT, Handler } from '~/components/containers/modal/PlaylistModal.vue';
 
 export type On = {
   [INPUT]: OnModal['input']
